Indices on "_latest" tables
This commit implements index tables that contain only data extracted from the `_latest` table of a repo, instead of indexing the full historical data and needing to join with `_latest` on each query. This requires deletion of old matching values from the index upon insertion.

This will improve performance of new indexes and in particular will allow "partial indices" to become truly partial (i.e. smaller than the tables they're based on). The change is "opt-in" with existing indices remaining unchanged.

Specifics:
- Added an `onLatest` parameter (defaulting to `false`) to `SecondaryIndex` and the APIs that create it.
- Added deletion logic for the new indices in table insertion and deletion methods.
- Added warnings when creating a multi-index or partial index that is _not_ based on a `_latest` table, since such an index will never delete values that are no longer matched.
- Added catch-up logic in `TableJanitor` for the new indices.
- Added optimizations to the retrieval methods that return only the index data, so that index tables they don't join with `_latest` tables for filtering when they're already based on them.
- Added `IndexTableSizeAction` to retrieve the "raw" table size of an index, without joining it with the `_latest` table. This is mostly useful for unit-testing the size of the new indices.

Note: the changes to the `IndexTableMethods` trait and the `SecondaryIndex` case class maintain source compatibility but break binary compatibility.