This commit upgrades several libraries needed to support Play 2.6 and Scala 2.12.
Libraries updated for Play 2.6 integration:
* Slick
* Twirl plugin

Libraries updated for Scala 2.12 integration:
* Scalatest
* Scalacheck
* Scala-async

Where a Repos dependency already had a higher version than the Play 2.6 dependency, I kept the repos version as-is. E.g. the Akka version here was left at 2.5.22, which is higher than the minimum 2.5.3 akka version in Play 2.6

This commit also adds the ability to cross compile with Scala 2.12.

Minor changes to the execution context are included with this commit.