Bumps minor project version.
We had previously thought that this project was auto-incremented upon releasing to Maven, (see comment on previous PR here: https://github.com/trueaccord/repos/pull/6#discussion_r365499239,) however, it appears that this project is manually versioned, released to an artifact repository, and then re-versioned as a snapshot (see commit log for examples: https://github.com/trueaccord/repos/commits/master)

Note: Bumping the minor version was chosen over bumping the patch version because in the time since that the last artifact was built we are additionally supporting Scala 2.12 via cross-building. The change is fully backwards compatible.